+++ order-management/backend/domain/order/model/request.go
@@ -2,5 +2,4 @@ package model
 
 import sharedOrderModel "tmossDev.github.com/eco-system/shared-components/backend/package/order/model"
 
-type OrderResponse = sharedOrderModel.OrderResponse
-type OrderItem = sharedOrderModel.OrderItem
+type UpdateOrderStatusRequest = sharedOrderModel.UpdateOrderStatusRequest